+++ cmd/som/cli/plot_error.go
@@ -12,6 +12,7 @@ func plotErrorCommand() *cobra.Command {
 	var size []int
 	var dataFile string
 	var labelsColumn string
+	var boundaries string
 	var delim string
 	var noData string
 	var rmse bool
@@ -46,7 +47,8 @@ For large datasets, --sample can be used to show only a sub-set of the data.`,
 
 			return plotHeatmap(size,
 				somFile, outFile, dataFile,
-				labelsColumn, delim, noData, title, ignore, sample,
+				labelsColumn, delim, noData, title,
+				ignore, boundaries, sample,
 				func(s *som.Som, p *som.Predictor, r table.Reader) (plotter.GridXYZ, []string, error) {
 					mse := p.GetError(rmse)
 					return &plot.FloatGrid{Size: *s.Size(), Values: mse}, nil, nil
@@ -56,6 +58,7 @@ For large datasets, --sample can be used to show only a sub-set of the data.`,
 	}
 
 	command.Flags().BoolVarP(&rmse, "rmse", "r", false, "Use root mean squared error instead of mean squared error")
+	command.Flags().StringVarP(&boundaries, "boundaries", "b", "", "Optional categorical variable to show boundaries for")
 
 	command.Flags().IntSliceVarP(&size, "size", "s", []int{600, 400}, "Size of the plot in pixels")
 	command.Flags().StringVarP(&dataFile, "data-file", "f", "", "Data file. Required")