+++ cmd/som/cli/plot_umatrix.go
@@ -12,6 +12,7 @@ func plotUMatrixCommand() *cobra.Command {
 	var size []int
 	var dataFile string
 	var labelsColumn string
+	var boundaries string
 	var delim string
 	var noData string
 	var ignore []string
@@ -36,7 +37,7 @@ For large datasets, --sample can be used to show only a sub-set of the data.`,
 			return plotHeatmap(size,
 				somFile, outFile, dataFile,
 				labelsColumn, delim, noData, "U-Matrix",
-				ignore, sample,
+				ignore, boundaries, sample,
 				func(s *som.Som, p *som.Predictor, r table.Reader) (plotter.GridXYZ, []string, error) {
 					uMatrix := s.UMatrix(true)
 					return &plot.UMatrixGrid{UMatrix: uMatrix}, nil, nil
@@ -45,6 +46,7 @@ For large datasets, --sample can be used to show only a sub-set of the data.`,
 		},
 	}
 
+	command.Flags().StringVarP(&boundaries, "boundaries", "b", "", "Optional categorical variable to show boundaries for")
 	command.Flags().IntSliceVarP(&size, "size", "s", []int{600, 400}, "Size of the plot in pixels")
 	command.Flags().StringVarP(&dataFile, "data-file", "f", "", "Data file. Required for --label")
 	command.Flags().StringSliceVarP(&ignore, "ignore", "i", []string{}, "Ignore these layers for BMU search")